refactor: rename module name - core.data.[datastore > settings], core.data.[retrofit > remote-db]
