+++ core/data/remote-db/src/main/AndroidManifest.xml
