+++ core/data/remote-db/src/main/java/com/example/core/data/remote_db/DbRemoteModule.kt
@@ -1,0 +1,2 @@
+package com.example.core.data.remote_db
+
