--- core/data/retrofit/src/main/java/com/example/core/data/retrofit/DbRemoteDataSource.kt
@@ -1,2 +1,0 @@
-package com.example.core.data.retrofit
-
