+++ core/data/settings/.gitignore
