+++ core/data/settings/build.gradle.kts
