+++ core/data/settings/proguard-rules.pro
