+++ core/data/settings/src/main/AndroidManifest.xml
