+++ settings.gradle.kts
@@ -25,11 +25,11 @@ include(":core:ui:ui")
 include(":core:ui:designsystem")
 include(":core:data:data")
 include(":core:model")
-include(":core:data:datastore")
+include(":core:data:settings")
 include(":feature:signin")
 include(":feature:workout")
 include(":feature:logs")
 include(":feature:more")
 include(":core:utils")
 include(":core:data:credentials")
-include(":core:data:retrofit")
+include(":core:data:remote-db")
